requirements.txt
primer3-py == 0.6.1